refactor(ui): move targets to seperate file
